+++ src/components/feature/CartSection/CartSection.styles.tsx
@@ -8,6 +8,10 @@ export const Container = styled.div`
   padding: 36px 24px 0;
 `;
 
+export const Description = styled.div`
+  font-size: 12px;
+`;
+
 export const CartList = styled.div`
   overflow: scroll;
   height: 400px;
@@ -21,4 +25,12 @@ export const CartList = styled.div`
 
   scrollbar-width: none;
   scrollbar-color: transparent transparent;
+`;
+
+export const EmptyCartContainer = styled.div`
+  text-size: 16px;
+  flex: 1;
+  display: flex;
+  justify-content: center;
+  align-items: center;
 `;
