+++ src/components/feature/CartSection/Header/index.tsx
@@ -1,16 +1,9 @@
 import * as S from "./Header.styles";
 
-const Header = ({
-  selectedCartItemCount,
-}: {
-  selectedCartItemCount: number;
-}) => {
+const Header = () => {
   return (
     <S.Container>
       <S.Title>장바구니</S.Title>
-      <S.Description>
-        현재 {selectedCartItemCount}종류의 상품이 담겨있습니다.
-      </S.Description>
     </S.Container>
   );
 };
